(356) Move fetchPlaces to services/ folder
